fix: beans ids, messagehandler to bean
