+++ bot/src/main/java/edu/java/bot/service/CommandHandler.java
@@ -14,4 +14,8 @@ public interface CommandHandler {
     }
 
     CommandName getCommand();
+
+    default String getBeanName() {
+        return getCommand().name();
+    }
 }
