+++ bot/src/main/java/edu/java/bot/service/ListHandler.java
@@ -10,7 +10,7 @@ import java.util.Arrays;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-@Component("LIST")
+@Component
 public class ListHandler implements CommandHandler {
     private final ApplicationConfig applicationConfig;
