+++ bot/src/main/java/edu/java/bot/service/StartHandler.java
@@ -11,7 +11,7 @@ import java.util.ArrayList;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-@Component("START")
+@Component
 public class StartHandler implements CommandHandler {
     private final ApplicationConfig applicationConfig;
