+++ bot/src/test/java/hw1/UserMessageHandlerTest.java
@@ -1,4 +1,4 @@
-package edu.java.hw1;
+package hw1;
 
 import com.pengrad.telegrambot.model.Chat;
 import com.pengrad.telegrambot.model.Message;
@@ -12,17 +12,19 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
+import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
-import static org.mockito.Mockito.mock;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+import static org.mockito.Mockito.mock;
 
 @ExtendWith(MockitoExtension.class)
 public class UserMessageHandlerTest {
     BotUser botUser = new BotUser(1, 123, "Test User", false);
     UserMessageHandlerImpl messageHandler = new UserMessageHandlerImpl();
-    Message message = mock(Message.class);
-    Update update = mock(Update.class);
+    @Mock Update update = new Update();
+
+    @Mock Message message = new Message();
     Chat chat = mock(Chat.class);
     User user = new User(1L);
 
@@ -54,7 +56,6 @@ public class UserMessageHandlerTest {
     @Test
     void extractsUser() {
         Mockito.when(update.message()).thenReturn(message);
-        Mockito.when(message.text()).thenReturn("/start");
         Mockito.when(message.chat()).thenReturn(chat);
         Mockito.when(message.chat().id()).thenReturn(1L);
         Mockito.when(message.from()).thenReturn(user);
